响应速度基本信息插入数据库；过深页面基本信息插入数据库

+++ src/main/java/com/trs/gov/kpi/entity/responsedata/PageDepthResponse.java
@@ -18,7 +18,7 @@ public class PageDepthResponse {
 
     private Integer pageDepth;
 
-    private Integer pageSpace;
+    private Long pageSpace;
 
     private Date checkTime;
 }
